pyproject.toml
[tool.poetry]
name = "airflow-junji"
version = "0.1.0"
description = "Proyecto para el manejo de librerias y scheduler de estas."
authors = ["Sebastian Navarro Saavedra <snavarro@junji.cl>"]
readme = "README.md"
packages = [{include = "airflow_junji"}]

[tool.poetry.dependencies]
python = "^3.11"
apache-airflow = "^2.6.1"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
pandas = "^2.0.2"
apache-airflow-providers-postgres = "^5.5.0"
office365-rest-python-client = "2.3.13"
office365 = "^0.3.15"
holidays = "^0.26"
apache-airflow-providers-sftp = "^4.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
bandit = "^1.7.5"
safety = "^2.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
